+++ CHANGELOG.md
@@ -5,6 +5,11 @@ PrestaShop 1.6.x, 1.7.x and 8.x
 
 # Changelog
 
+## [3.8.3]
+### Fixed
+- Capture the payment only from the reserved amount on status update.
+- Fixed the Smarty template issue when redirecting to the success page.
+
 ## [3.8.2]
 ### Fixed
 - Load plugin javascript only on Checkout page.